qtwebkit: drop obsolete patch 0001-jsc-Add-licuuc-licui18n-LIBS.patch
This patch caused a linker error:

| linking ../../bin/jsc
| [...]ld: cannot find -licuuc -licui18n

The linker command-line contains:

| [...] -licuuc\ -licui18n [...] -licui18n -licuuc -licudata [...]

So the problem gets caused by the backslash, making the actual library
name "libicuuc -licui18n.so". This may have been caused be the double
quotes used in the patch. However, since both libs are already part
of the command-line, the patch can just get dropped. The patch indicates,
that it was created for a beta version, so I assume that this was
fixed before the 5.2 release.

Signed-off-by: Andreas Oberritter <obi@opendreambox.org>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>