qtbase.inc: PACKAGECONFIG[gstreamer] gst-plugins-base dependency missing
When testing for GStreamer support, the qtbase 5.1.1 configure script
tries to determine pkg-config information for both gstreamer-0.10 and
gstreamer-plugins-base-0.10.

If either one is unavailable then the configure test fails (with the
slightly confusing error that gst/gst.h can't be found, even when the
gstreamer header files are present in sysroot).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>