nativesdk-qtbase-tools: disable Debian-style package renaming
Since commit 8da5017712a7ddcb8488e095b9aa03b0273f23e9, libQt5*.so files
are included in the nativesdk-qtbase-tools package. However, this together
with the fact that the binaries are placed in a subdir "qt5" of bin will
confuse the Debian-style package renaming logic in debian.bbclass, causing the
resulting ipk package to be created as "nativesdk-libqt5core5".

Signed-off-by: Jacob Kroon <jacob.kroon@mikrodidakt.se>
Acked-by: Otavio Salvador <otavio@ossystems.com.br>