qt5.inc: Fix packaging of architecture independent data
The architecture independent data were being added to the '-dev'
package however it shouldn't be. For example the QtWebEngine had the
translation and the resources package were not being installed with
the main package, as it should be.

This commit changes the qttools and qttranslations recipes as well so
those keep its contents in the respective packages:

 qttools: add the phrasebooks in the qttools-tools package
 qttranslations: add a new subpackage for qtquickcontrols

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>