qtconnectivity, qtsystems: fix bluetooth support
QMAKE_CACHE_EVAL is only used in qtwebengine and qtwebkit, but not in
the rest of the QT5 packages, so use EXTRA_QMAKEVARS_PRE to pass the
packageconfig options to the do_configure task.

Also generalize bluetooth support so it can be used not only with bluez4
but with bluez5 as well.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>