qtbase: fix up pkgconfig replacements
The replacements were being done in qt_module.prf which is not included
when building 3rd party libraries. To resolve this, move the
replacements from qt_module.prf to qt.prf which is included when
building any qmake project. The pkgconfig include replacement is removed
as it doesn't seem to be needed anymore and causes issues when the
includedir path in pkgconfig has part of its path replaced with
${includedir} resulting in a circular reference that causes the OE
package_do_pkgconfig task to fail.

Signed-off-by: Jonathan Liu <net147@gmail.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>