qt5: Allow disabling SHA validation for branch
Qt releases are done from a release branch (5.x.x), which is later merged
back to a stable branch (5.x) and then removed. When a recipe is updated
to use the released SHA from the release branch, it will eventually break
when the release branch is removed. This happens because bitbake tries to
validate that given SHA is found from the given branch. Add additional
variable that can be used to disable the SHA check when branch is known
to be later removed, but SHA remains valid.

Signed-off-by: Samuli Piippo <samuli.piippo@qt.io>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>