qtbase: add patch to make Qt5Config modules behave better with cmake builds
Currently if a project if using the cmake bbclass and the CMakeLists.txt file
uses find_package to find Qt5 and error like this is produced:

|   The imported target "Qt5::Core" references the file
|
|      "/qmake"
|
|   but this file does not exist.  Possible reasons include:

This is because OE_QMAKE_PATH_EXTERNAL_HOST_BINS is only set in the cmake_qt5 bblcass.

To make it work better in that case the Qt5Config files check if OE_QMAKE_PATH_EXTERNAL_HOST_BINS
is defined and bail out if not. For CMake this is the same as Qt5 wasn't found.

Signed-off-by: Pascal Bach <pascal.bach@siemens.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>