qt5: update to Qt 5.8
The linux-oe-g++ mkspec is changed to use $$(...) operator to obtain
the contents of an environment value when qmake is run instead of when
Makefile is processed. All OE_QMAKE_xxx variables need to be exported
for qmake to find them. configure's setBootstrapVariable function needs
to change $$(..) to normal $(...) operator to work with qmake's Makefile.

qt.conf generation for qtbase recipes is not needed, as configure will
generate its own version based on configure arguments. Skip running
qmake, since configure is now automatically invoked when it's run in
qtbase's root folder.

Change-Id: I4c937c1acb3d204ac4560090cab3d0fc7e0e815a
Reviewed-by: Risto Avila <risto.avila@qt.io>
Reviewed-by: Jonathan Liu <net147@gmail.com>
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>