qtbase: remove absolute binary paths from mkspec files
When running 'bitbake -c populate_sdk' on an image that has qt5webkit
enabled, the uikit feature is enabled in qtbase, which then fails to install
with

  error: qtbase-mkspecs-5.8.0+git0+49dc9aa409-r0 conflicts with /usr/bin/perl

The culprit for this is /usr/lib/qt5/mkspecs/features/uikit/devices.pl,
which has #!/usr/bin/perl hardcoded. Fix this in a similar way other recipes
do it as well and strip out these bits at install time.

Signed-off-by: Daniel Mack <daniel@zonque.org>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>