gstreamer1.0-plugins-bad: add qt5 packageconfig
This adds a new command line switch to the configure script to make sure
the right paths to moc, uic, rcc are used. It also makes sure OpenGL is
enabled if the qt5 packageconfig is in use.

Note that the moc/uic/rcc path configure switches were introduced
in GStreamer 1.12.2, so this version is a requirement.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>