qtbase-native: Force use of built-in pcre
If we specify neither -system-pcre nor -qt-pcre then qtbase will first see
if it can find the pcre library in the system and fall back to compiling
its own if that's not possible.

Unfortunately, this means that if qtbase-native happens to be built on a
machine with system pcre available then the resulting binaries such as
bin/qt5/uic in sstate files won't work correctly on machines that don't
have pcre available. Even ignoring that, it doesn't help for packages to
build differently depending on how the build system is configured.

We could depend on pcre-native and pass -system-pcre, but this doesn't work
by default because qtbase actually requires pcre16 which is an optional
(via PACKAGECONFIG) feature of pcre so the default there would need to be
changed too.

So, let's make qtbase-native match the default of qtbase by passing
-qt-pcre in order to ensure that qtbase-native is always built with its
own pcre implementation.

Signed-off-by: Mike Crowe <mac@mcrowe.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>