qtbase: mkspecs: add default toolchain options
While most mkspecs use hardcoded compiler names / options, qmake in
meta-qt5 is setup to get them from the environment. While this works
fine with openembedded, it does break normal workflows outside of it.
For example to get the SDK working with Qt Creator either the environmental
script most be sourced first, making it impossible to switch between
different Qt versions or the OE_QMAKE must be set for every project,
otherwise you get cryptic errors like 'make: c: Command not found',
since the compiler name is not set and hence an attempt is made to
execute only the arguments.

Likewise on the target itself, qmake doesn't work without first setting
the OE_QMAKE variables.

This patch adds default toolchains options, so it is no longer required
to set them, but keeps the flexibility to alter them from the environment.

Signed-off-by: Jeroen Hofstee <jhofstee@victronenergy.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>