qt5: add option to build Qt5 statically
Add PACKAGECONFIG that enables static build for qtbase and consequently
to all Qt recipes. This can be enabled with DISTRO_FEATURE 'qt5-static',
which also adds qtdeclarative-native dependency to all recipes that
depend on qtdeclarative as it's required for qmlimportscannertool.

Building examples and tests with static build can take excessive amounts
of time and disk space, so disabling DISTRO_FEATURE 'ptest' and
PACKAGECONFIGs 'examples' from qtbase is advised.

Not all recipes support static builds and those are not fixed here.

Change-Id: Ia0e2a2467cd42d4395ed5292a645fd1d89ad521a
Signed-off-by: Samuli Piippo <samuli.piippo@qt.io>