gstreamer1.0-plugins-good: fix qt5 dependency
Commit 73f99f2 ("gstreamer1.0-plugins-{good,bad}: move the qt5
PACKAGECONFIG from bad to good bbappend") just copied the bbappend file
without checking the dependencies.

The qt5 plugin needs the gl plugin which was moved in gstreamer 1.14
from bad to base. So the pyhton() function checking the PACKAGECONFIG_GL
isn't needed here, furthermore it causes bitbake errors:

     0013:        gl_packageconfig = d.getVar('PACKAGECONFIG_GL', expand=True)
     0014:        with open("/tmp/marcotest", "w") as f:
     0015:            print(gl_packageconfig, file=f)
     0016:        d.appendVar('PACKAGECONFIG', ' ' + gl_packageconfig)
 *** 0017:}
Exception: TypeError: can only concatenate str (not "NoneType") to str

The patch removes the python() function to fix this and adds the
gstreamer1.0-plugins-base dependency to have access to the gl plugin.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>