qtbase: Export host compiler and linker
Certain portions of qtwebengine build involves building native tools
using host compiler/linker, this however currently uses gcc and when we
build with clang to compile the target pieces, chromium configures for
clang and rightly so, but then does not have a way to define flags for
host and target separately, and it ends up supplying same flags for both
host and target builds, the trouble is that the options between clang
and gcc are not always same in some cases, they dont understand each
others options. So build starts to fail due to these errors about
unknown cmdline options.

This makes sure we inject CC_host, CXX_host, LD_host into qmake.conf so
they can be used in such packages

We use clang to do host builds when TOOLCHAIN = "clang" is used
otherwise default to gcc as it was the case

Signed-off-by: Khem Raj <raj.khem@gmail.com>