qtbase-native: add native libdir to search path
After commit 521a85395, Qt uses global search paths to resolve each library
path, but because we are using -no-gcc-sysroot the native libdir is not part
of the search path. Add the native libdir to search path so that we don't
depend on host libraries for locating zlib.

Fixes "ERROR: Feature 'system-zlib' was enabled, but the pre-condition
'libs.zlib' failed."

Change-Id: I25ec41ee26e7c9406ba0107a766258532b1c8d90
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>