qtbase, packagegroup-qt5-toolchain-target: widgets are needed by qtcharts
qtcharts depends on widgets in PACKAGECONFIG of qtbase.
Right now this dependency is not checked in RDEPENDS of
packagegroup-qt5-toolchain-target.

If widgets are removed from PACKAGECONFIG of qtbase the error occurs,
that nothing provides qtcharts-dev / qtcharts-mkspecs / qtcharts-qmlplugins.

With this patch qtcharts-* will only be in RDEPENDS, if widgets are set in
PACKAGECONFIG of qtbase. Therefore widgets need to be added in DISTRO_FEATURES.

Signed-off-by: Norbert Wesp <n.wesp@phytec.de>