cmake-nativesdk: install toolchain file for Qt5 bins.
This installs a cmake file setting the OE_QMAKE_PATH_EXTERNAL_HOST_BINS
variable to the location of the native Qt5 binaries in the SDK as given
by the OE_QMAKE_PATH_HOST_BINS environment variable. This is necessary
to be able to use the SDK to build Qt-applications for the target. See
also 05103827d50ec which introduced the error message intended for
incorrect usage of the cmake bbclass rather than the cmake_qt5 bbclass.

Signed-off-by: Bjørn Øivind Bjørnsen <bo.bjornsen@gmail.com>