Revert "recipes: Use features_check instead of distro_features_check"
This reverts commit b7f4bd209c03c4f765cec84f4aa247819a6bb2f3.

I've accidentally merged this to zeus which doesn't have the feautres_check.bbclass
yet.

Even for master branch it's problematic, because some people tend to use different
branch of meta-qt5 with different OE releases, so if you need to use meta-qt5/master
with zeus or older, you'll need to introduce features_check.bbclass (with just
"inherit distro_features_check" in your own layer for the compatibility.

In the oposite case when using meta-qt5/zeus (or older) with master/dunfell, you'll
just get bunch of warnings about deprecated distro_features_check inherit which is
safe to ignore.