qtbase: eglfs: Prevent printing a warning in every widget-based app
Attempting to reflect reality with the value of AA_ShareOpenGLContexts
was a valiant attempt, but it is not possible to do setAttribute without
getting a warning now once QGuiApplication is constructed. So just leave
the attribute as is, changing its value is not strictly required.

This fix is backported from qtbase upstream, it fixes: QTBUG-85796.

Signed-off-by: Ming Liu <liu.ming50@gmail.com>