qmake5: Move to EXPORT_FUNCTIONS
A recent change in bitbake [1] has caused a build failure in qtwebkit.
The fix with how the EXPORT_FUNCTIONS is being handled means that it
is running into a collision with the explicit do_configure defined in this
file.

This fix moves the explicit do_configure and do_insatll to use the
EXPORT_FUNCTIONS methodology so that the conflict is resolved correctly.

[1] https://git.openembedded.org/bitbake/commit/?id=66306d5151acb0a26a171c338d8f60eb9eb16c6b

Signed-off-by: Ryan Eatmon <reatmon@ti.com>