qtbase: packageconfig: fix gl in response to upstream libglvnd updates
In libglvnd on whinlatter, virtual/libgl is now only provided when the
distro has x11. This is as it should be, but now configurations that
use wayland need to not require virtual/libgl since they can't have it
anymore. PACKAGECONFIG is updated for the non-x11 graphics case to
require the things it needs instead (kms, gbm, gles2, eglfs).

Signed-off-by: Kurt Kiefer <kekiefer@gmail.com>
Signed-off-by: Gaël PORTAY <gael.portay@rtone.fr>
(cherry picked from commit 0c9f841c4bd9049c1037678474eedee63e669599)